@@ -3,8 +3,8 @@
   "version": "0.1.0",
   "license": "MIT",
   "engines": {
-    "node": "8.6.0",
-    "npm": "5.4.2"
+    "node": "8.9.1",
+    "npm": "5.5.1"
   },
   "scripts": {
     "ng": "ng",
@@ -16,36 +16,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "4.4.6",
-    "@angular/compiler": "4.4.6",
-    "@angular/core": "4.4.6",
-    "@angular/forms": "4.4.6",
-    "@angular/http": "4.4.6",
-    "@angular/platform-browser": "4.4.6",
-    "@angular/platform-browser-dynamic": "4.4.6",
-    "@angular/router": "4.4.6",
+    "@angular/common": "5.1.0",
+    "@angular/compiler": "5.1.0",
+    "@angular/core": "5.1.0",
+    "@angular/forms": "5.1.0",
+    "@angular/http": "5.1.0",
+    "@angular/platform-browser": "5.1.0",
+    "@angular/platform-browser-dynamic": "5.1.0",
+    "@angular/router": "5.1.0",
     "core-js": "2.5.1",
     "hint.css": "2.5.0",
-    "rxjs": "5.5.0",
+    "rxjs": "5.5.5",
     "zone.js": "0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "1.4.9",
-    "@angular/compiler-cli": "4.4.6",
-    "@types/jasmine": "2.6.0",
-    "@types/node": "8.0.46",
-    "codelyzer": "3.2.1",
+    "@angular/cli": "1.6.0",
+    "@angular/compiler-cli": "5.1.0",
+    "@types/jasmine": "2.8.2",
+    "@types/node": "8.0.57",
+    "codelyzer": "4.0.1",
     "jasmine-core": "2.8.0",
     "jasmine-spec-reporter": "4.2.1",
     "karma": "1.7.1",
     "karma-chrome-launcher": "2.2.0",
     "karma-cli": "1.0.1",
     "karma-coverage-istanbul-reporter": "1.3.0",
-    "karma-jasmine": "1.1.0",
+    "karma-jasmine": "1.1.1",
     "karma-jasmine-html-reporter": "0.2.2",
-    "protractor": "5.2.0",
+    "protractor": "5.2.1",
     "ts-node": "3.3.0",
-    "tslint": "5.7.0",
-    "typescript": "2.5.3"
+    "tslint": "5.8.0",
+    "typescript": "2.6.2"
   }
 }
